# [Tree] Improve Tree Performance by replacing computed height with TreeItem's cached minimum size