--- a/scene/resources/text_paragraph.h
+++ b/scene/resources/text_paragraph.h
@@ -152,6 +152,8 @@ public:
 
 	int hit_test(const Point2 &p_coords) const;
 
+	bool is_dirty();
+
 	Mutex &get_mutex() const { return _thread_safe_; }
 
 	TextParagraph(const String &p_text, const Ref<Font> &p_font, int p_font_size, const String &p_language = "", float p_width = -1.f, TextServer::Direction p_direction = TextServer::DIRECTION_AUTO, TextServer::Orientation p_orientation = TextServer::ORIENTATION_HORIZONTAL);