Originally, `TreeItem`'s content height was computed in every draw while traversing the tree, even if the item content never changes. And there's already `minimum_size` cached in each item, which is computed in a similar way.

Tested on a massive plain-tree with 15K and 150K items(project file attached below), and got significant performance improvement (About 5~10x).
Here're results on my desktop when scrolling to near the bottom of tree.
Amount | This PR | Master
---------| ---------| --------
15K Items | 1543FPS | 152FPS
150K Items | 36FPS | 3FPS, and with a lot much longer start time

I also checked the behaviors of more complex `Tree`s in the editor, and didn't notice any obvious artifacts, but I'm not very sure if there were any side-effect after changing these.

Might partially fix these issues?
* *Bugsquad edit, closes: https://github.com/godotengine/godot/issues/70869*
* *Bugsquad edit, closes: https://github.com/godotengine/godot/issues/83460*

EDIT:
After some "extreme" test with huge amount of auto-wrapping text `TreeItems` and scrolling fast will cause some item's height cache being outdated. 
~~I'm still looking for some other way to make this work. So convert this to draft now and will try reopen when I figured out.~~
I believe that's not related to this PR's modification, as I can reproduce the above issue in both 4.3 and master branch.  I'll open another issue about that.


[treetest.zip](https://github.com/user-attachments/files/16959026/treetest.zip)