@scgm0 2024-09-12
After testing, it works great in my personal project, now I can directly use `Tree` to build my tools 👍