@KoBeWi 2024-09-13
I see some discrepancy between the code you deleted in `compute_item_height()` and TreeItem's `get_minimum_size()`, e.g. the latter does not account for `theme_cache.checked` height.