@maidopi-usagi 2024-09-14
> I see some discrepancy between the code you deleted in `compute_item_height()` and TreeItem's `get_minimum_size()`, e.g. the latter does not account for `theme_cache.checked` height.

I'm wondering the reason that `get_minimum_size()` didn't take `theme_cache.checked`'s height into account. If there's nothing affected I'll re-add that into the method.

Also I noticed that `update_item_cell()` seems to be called twice every time (another in `get_minimum_size()`). 
Maybe we can remove the following one?

https://github.com/godotengine/godot/blob/f5f813e9674fa433a9702ea614bcd56e59dd80db/scene/gui/tree.cpp#L1824-L1827