@KoBeWi 2024-10-05
> Also I noticed that update_item_cell() seems to be called twice every time (another in get_minimum_size()).

Only once call will be effective, unless the cell becomes dirty again for some reason.