@KoBeWi 2024-10-05
The scene tree is noticeably bigger vertically. The left is with this PR, the right is without:
![image](https://github.com/user-attachments/assets/616b052c-95e9-49b1-8b19-942c791236c7)