@maidopi-usagi 2024-10-06
> The scene tree is noticeably bigger vertically. The left is with this PR, the right is without: ![image](https://private-user-images.githubusercontent.com/2223172/373912138-616b052c-95e9-49b1-8b19-942c791236c7.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjgyMTI4NjgsIm5iZiI6MTcyODIxMjU2OCwicGF0aCI6Ii8yMjIzMTcyLzM3MzkxMjEzOC02MTZiMDUyYy05NWU5LTQ5YjEtOGIxOS05NDJjNzkxMjM2YzcucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI0MTAwNiUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNDEwMDZUMTEwMjQ4WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9NjhkZDgxZTQ3ZjlhZTkyNDk2MmFjYjk0NmYyYTBiNWM2MmYyODE0ZGRhZDIyMTM4NzEwOWYyMmE2MzVlODg2MiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.8kgx6DnTljD2-GnKhG0ELq-S2xTS3iSOKhOStmHssqw)

Thanks for pointing out! 
I've found that's because of the cached minimum size computes buttons height using the combination of button texture's size and `button_pressed`'s minimum_size, compared with the old calculation method. 
https://github.com/godotengine/godot/blob/f5f813e9674fa433a9702ea614bcd56e59dd80db/scene/gui/tree.cpp#L1542-L1552