@KoBeWi 2024-11-06
> Or should it use the maximum height instead?

There is some weird behavior when `button_pressed` is bigger than other styles, but it's the same as the one on master, so I think it's fine.